+++ bundles/org.enterprisedomain.classmaker/src/org/enterprisedomain/classmaker/jobs/export/PDEPluginExporter.java
@@ -53,7 +53,7 @@ import org.osgi.framework.Version;
 @SuppressWarnings("restriction")
 public class PDEPluginExporter extends AbstractExporter {
 
-	public PDEPluginExporter(int stateTimestamp) {
+	public PDEPluginExporter(long stateTimestamp) {
 		super(stateTimestamp);
 	}
